AI PROJECT NEW/requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
requests==2.31.0
python-dotenv==1.0.1
jira